refactor: convert services to use SQLAlchemy repositories and remove in-memory logic

+++ app/services/project_service.py
@@ -1,23 +1,19 @@
-import uuid
 from typing import List
 from app.models.project import Project
-from app.repositories.project_repository import IProjectRepository
-from app.repositories.task_repository import ITaskRepository
 from app.models.task import Task
+from app.repositories.project_sql_repository import SQLAlchemyProjectRepository
+from app.repositories.task_sql_repository import SQLAlchemyTaskRepository
 
 class ProjectService:
     """Business logic for projects."""
     
-    def __init__(self, project_repo: IProjectRepository, task_repo: ITaskRepository):
+    def __init__(self, project_repo: SQLAlchemyProjectRepository, task_repo: SQLAlchemyTaskRepository):
         self.project_repo = project_repo
         self.task_repo = task_repo
-        self.all_projects_counter = 0
-        
+
     def create_project(self, name: str, description: str) -> Project:
-        project_id = self.all_projects_counter + 1
-        project = Project(id=project_id, name=name, description=description)
+        project = Project(name=name, description=description)
         self.project_repo.save(project)
-        self.all_projects_counter+=1
         return project
 
     def get_all_projects(self) -> List[Project]:
@@ -32,19 +28,14 @@ class ProjectService:
     def delete_project(self, project_id: str) -> None:
         self.project_repo.delete(project_id)
 
-    def add_task_to_project(
-        self, project_id: str, title: str, description: str, deadline: str
-    ) -> Task:
+    def add_task_to_project(self, project_id: str, title: str, description: str, deadline: str) -> Task:
         project = self.get_project(project_id)
-        task_id = project.get_project_tasks_counter() + 1
-        task = Task(id=task_id, title=title, description=description, deadline=deadline)
-        project.add_task(task)
+        task = Task(title=title, description=description, deadline=deadline, project_id=project.id)
         self.task_repo.save(task)
-        self.project_repo.save(project)  # update project
         return task
 
     def remove_task_from_project(self, project_id: str, task_id: str) -> None:
-        project = self.get_project(project_id)
-        project.remove_task(task_id)
+        task = self.task_repo.get_by_id(task_id)
+        if not task or task.project_id != int(project_id):
+            raise ValueError("Task not found in this project")
         self.task_repo.delete(task_id)
-        self.project_repo.save(project)  # update project
